add skip function.

+++ python/lryzx.py
@@ -1,4 +1,5 @@
 #!/usr/bin/python
+import os
 import sys
 import random
 import math
@@ -32,16 +33,36 @@ if len(sys.argv) < 3:
     print 'Usage: train.yzx.txt test.yzx.txt'
     exit(-1)
 
+if not os.path.exists(sys.argv[2] + '.lr.weight'):
+    for round in range(0, trainRounds):
+        # train for this round
+        fi = open(sys.argv[1], 'r')
+        lineNum = 0
+        trainData = []
+        for line in fi:
+            lineNum = (lineNum + 1) % bufferCaseNum
+            trainData.append(ints(line.replace(":1", "").split()))
+            if lineNum == 0:
+                for data in trainData:
+                    clk = data[0]
+                    mp = data[1]
+                    fsid = 2 # feature start id
+                    # predict
+                    pred = 0.0
+                    for i in range(fsid, len(data)):
+                        feat = data[i]
+                        if feat not in featWeight:
+                            featWeight[feat] = nextInitWeight()
+                        pred += featWeight[feat]
+                    pred = sigmoid(pred)
+                    # start to update weight
+                    # w_i = w_i + learning_rate * [ (y - p) * x_i - lamb * w_i ] 
+                    for i in range(fsid, len(data)):
+                        feat = data[i]
+                        featWeight[feat] = featWeight[feat] * (1 - lamb) + eta * (clk - pred)
+                trainData = []
 
-for round in range(0, trainRounds):
-    # train for this round
-    fi = open(sys.argv[1], 'r')
-    lineNum = 0
-    trainData = []
-    for line in fi:
-        lineNum = (lineNum + 1) % bufferCaseNum
-        trainData.append(ints(line.replace(":1", "").split()))
-        if lineNum == 0:
+        if len(trainData) > 0:
             for data in trainData:
                 clk = data[0]
                 mp = data[1]
@@ -55,77 +76,58 @@ for round in range(0, trainRounds):
                     pred += featWeight[feat]
                 pred = sigmoid(pred)
                 # start to update weight
-                # w_i = w_i + learning_rate * [ (y - p) * x_i - lamb * w_i ] 
+                # w_i = w_i + learning_rate * [ (y - p) * x_i - lamb * w_i ]
                 for i in range(fsid, len(data)):
                     feat = data[i]
                     featWeight[feat] = featWeight[feat] * (1 - lamb) + eta * (clk - pred)
-            trainData = []
+        fi.close()
 
-    if len(trainData) > 0:
-        for data in trainData:
+        # test for this round
+        y = []
+        yp = []
+        fi = open(sys.argv[2], 'r')
+        for line in fi:
+            data = ints(line.replace(":1", "").split())
             clk = data[0]
             mp = data[1]
             fsid = 2 # feature start id
-            # predict
             pred = 0.0
             for i in range(fsid, len(data)):
                 feat = data[i]
-                if feat not in featWeight:
-                    featWeight[feat] = nextInitWeight()
-                pred += featWeight[feat]
+                if feat in featWeight:
+                    pred += featWeight[feat]
             pred = sigmoid(pred)
-            # start to update weight
-            # w_i = w_i + learning_rate * [ (y - p) * x_i - lamb * w_i ]
-            for i in range(fsid, len(data)):
-                feat = data[i]
-                featWeight[feat] = featWeight[feat] * (1 - lamb) + eta * (clk - pred)
-    fi.close()
+            y.append(clk)
+            yp.append(pred)
+        fi.close()
+        auc = roc_auc_score(y, yp)
+        rmse = math.sqrt(mean_squared_error(y, yp))
+        print str(round) + '\t' + str(auc) + '\t' + str(rmse)
 
-    # test for this round
-    y = []
-    yp = []
+    # output the weights
+    fo = open(sys.argv[1] + '.lr.weight', 'w')
+    featvalue = sorted(featWeight.iteritems(), key=operator.itemgetter(0))
+    for fv in featvalue:
+        fo.write(str(fv[0]) + '\t' + str(fv[1]) + '\n')
+    fo.close()
+
+
+# output the prediction
+if not os.path.exists(sys.argv[2] + '.lr.pred'):
     fi = open(sys.argv[2], 'r')
+    fo = open(sys.argv[2] + '.lr.pred', 'w')
+
     for line in fi:
         data = ints(line.replace(":1", "").split())
-        clk = data[0]
-        mp = data[1]
-        fsid = 2 # feature start id
         pred = 0.0
-        for i in range(fsid, len(data)):
+        for i in range(1, len(data)):
             feat = data[i]
             if feat in featWeight:
                 pred += featWeight[feat]
         pred = sigmoid(pred)
-        y.append(clk)
-        yp.append(pred)
+        fo.write(str(pred) + '\n')    
+    fo.close()
     fi.close()
-    auc = roc_auc_score(y, yp)
-    rmse = math.sqrt(mean_squared_error(y, yp))
-    print str(round) + '\t' + str(auc) + '\t' + str(rmse)
-
-# output the weights
-fo = open(sys.argv[1] + '.lr.weight', 'w')
-featvalue = sorted(featWeight.iteritems(), key=operator.itemgetter(0))
-for fv in featvalue:
-    fo.write(str(fv[0]) + '\t' + str(fv[1]) + '\n')
-fo.close()
-
-
-# output the prediction
-fi = open(sys.argv[2], 'r')
-fo = open(sys.argv[2] + '.lr.pred', 'w')
-
-for line in fi:
-    data = ints(line.replace(":1", "").split())
-    pred = 0.0
-    for i in range(1, len(data)):
-        feat = data[i]
-        if feat in featWeight:
-            pred += featWeight[feat]
-    pred = sigmoid(pred)
-    fo.write(str(pred) + '\n')    
-fo.close()
-fi.close()
